+++ packages/backend/src/postgre.ts
@@ -69,6 +69,7 @@ import { UserPublickey } from '@/models/entities/UserPublickey.js';
 import { UserSecurityKey } from '@/models/entities/UserSecurityKey.js';
 import { Webhook } from '@/models/entities/Webhook.js';
 import { Channel } from '@/models/entities/Channel.js';
+import { RetentionAggregation } from '@/models/entities/RetentionAggregation.js';
 
 import { Config } from '@/config.js';
 import MisskeyLogger from '@/logger.js';
@@ -182,6 +183,7 @@ export const entities = [
 	UserPending,
 	Webhook,
 	UserIp,
+	RetentionAggregation,
 	...charts,
 ];